requirements.txt
psycopg2
kafka-python
pyspark==3.4.2
